livefront-backend/requirements.txt
# fastapi==0.111.0
# uvicorn==0.29.0
# sse-starlette==1.6.1
# python-dotenv==1.0.1
# langchain==0.2.0
# langchain-community==0.2.0
# transformers==4.41.1
# torch==2.2.2
# accelerate==0.29.3
# SQLAlchemy==2.0.30
# python-multipart==0.0.9
# pydantic_settings==2.0.4
# fastapi
# uvicorn[standard]
# python-dotenv
# openai
# faiss-cpu
# numpy
# loguru
pydantic-settings
fastapi
uvicorn[standard]
python-dotenv
loguru
transformers
torch
sentence-transformers
faiss-cpu
sqlalchemy
aiosqlite
greenlet